Modularized security configuration and introduced DPoP support.
